+++ changelog.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## 2026-08-12
+- [x] [fix discounted Full Database Stripe webhook fulfillment](docs/changelog/2026-08-12-discounted-full-database-webhook-fulfillment.md)
+
 ## 2026-08-11
 - [x] [fix Stripe webhook processing for unsuccessful Checkout Sessions](docs/changelog/2026-08-11-stripe-webhook-failure-processing.md)